Convert transition.{js,ts}
