+++ parseConfig.js
@@ -128,7 +128,7 @@ export default async function parseConfig({ defaults: defaultsIn = {}, clips, ar
     const userClipDurationOrDefault = userClipDuration || defaults.duration;
     if (videoLayers.length === 0) assert(userClipDurationOrDefault, `Duration parameter is required for videoless clip ${clipIndex}`);
 
-    const transition = calcTransition(defaults, userTransition, clipIndex === clips.length - 1);
+    const transition = calcTransition(defaults.transition, userTransition, clipIndex === clips.length - 1);
 
     let layersOut = flatMap(await pMap(layers, async (layerIn) => {
       const globalLayerDefaults = defaults.layer || {};
